+++ src/rollup.ts
@@ -1,3 +1,3 @@
-import preactIslandPlugin from './index.js'
+import preactIslandPlugin from './plugin'
 
 export default preactIslandPlugin.rollup
